Test Cases, Frontend re-deployment, Changing file path/names to be more clear
This push has a lot of changes at first glance most of it is changing the file path as it relates to the names of directories. I recreated the front end correctly as well using components to help create a navbar and pages to that can be viewed. I also contributed to the test cases by running a complete user CRUD.